requirements.txt
django
psycopg2
djangorestframework
django-filter